updated basic classes structure
