+++ ProjectSyncUpServiceLambda/src/main/java/com/nashss/se/projectsyncup/dependency/ServiceComponent.java
@@ -1,12 +1,8 @@
 package com.nashss.se.projectsyncup.dependency;
 
 import com.amazonaws.services.lambda.runtime.RequestHandler;
-import com.nashss.se.projectsyncup.activity.AddTaskToProjectActivity;
-import com.nashss.se.projectsyncup.activity.CreateProjectActivity;
+import com.nashss.se.projectsyncup.activity.*;
 
-import com.nashss.se.projectsyncup.activity.GetProjectActivity;
-
-import com.nashss.se.projectsyncup.activity.GetProjectTasksActivity;
 import com.nashss.se.projectsyncup.activity.requests.GetProjectTasksRequest;
 import dagger.Component;
 
@@ -43,4 +39,10 @@ public interface ServiceComponent {
      * @return GetProjectTasksActivity
      */
     GetProjectTasksActivity provideGetProjectTasksActivity();
+
+    /**
+     * Provides the relevant activity.
+     * @return RemoveTaskFromProjectActivity
+     */
+    RemoveTaskFromProjectActivity provideRemoveTaskFromProjectActivity();
 }
